requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
